requirements.txt
asgiref==3.7.2
blinker==1.6.2
boto3==1.28.33
botocore==1.31.33
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cryptography==41.0.3
cycler==0.10.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.42.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.4
s3transfer==0.6.2
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.19
supervision==0.13.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
wget==3.2
